Geolocation is saved in database for each FoodTruck account
Used MKLocalSearchCompleter to create queries for the search which returned a string address of the location. Used geoCodeAddressString to convert this string to a CLPlacemark which allowed me to have access to latitude and longitude for an address. Finally, used geoPointWithLatitude to convert the values to a PFGeoPoint and saved it to my backend