Refactored to use MKMarkerAnnotation. Added clustering identifier.
I decided to switch to a marker annotation because I wanted the title to be visible for each marker during all instances. Using a normal pin annotation or viewAnnotation would not allow this to happen unless I set the selected value to be true for all rendered pins. Therefore, it made sense to switch to markers because titles are rendered by default and are used as the standard marker in iOS 14