Users can now compose a review and view all reviews for a truck
Implemented scrolling collection view on details page so users can view all reviews for a given truck. Also added a compose VC that has a max character limit of 200 for writing reviews. This works with HCSStarRatingView and allows the user to use an interactable component to give a star rating